jourmungandr: move all tests class in journey_comment_test file
